+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 IMPROVEMENT
 - [docs] Include `ecosystem.json` and `tendermint-bft.md` from deprecated `aib-data` repository.
 
+IMPROVEMENTS:
+- [config] Add `instrumentation.max_open_connections`, which limits the number
+  of requests in flight to Prometheus server (if enabled). Default: 3.
+
 ## 0.22.0
 
 *July 2nd, 2018*